serious effeciency improvements
we were not using the indexed version of the matchqueen simple table
query in the update function. this function gets called a lot. this made
a huge difference.
also we were drawing the same hex over and over again , because of the
find neighbors call. this also gets called alot. we were also calling
update twice for every hex. cut it down to one.
also some bug fixes.